Voice Agent/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
twilio
pydantic
jinja2
